+++ Assets/Scripts/Tower/TowerBehaviour/TowerAttackBehaviour/TowerAttackBehaviourDamage.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class TowerAttackBehaviourDamage : TowerAttackBehaviour
 {
-    public override void Attack(Entity target, int impact)
+    public override void Attack(Entity target, float impact)
     {
         target.GetDamage(impact);
         //Debug.Log($"Entity receives {impact} damage! Entities remaining health: {target.GetHP()}");
